Issue #11163: Split InputHiddenField2 into focused test inputs
