+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/coding/HiddenFieldCheckTest.java
@@ -149,47 +149,86 @@ public class HiddenFieldCheckTest
     }
 
     @Test
-    public void testDefault()
-            throws Exception {
+    public void testHiddenField2Basic() throws Exception {
         final String[] expected = {
-            "30:13: " + getCheckMessage(MSG_KEY, "hidden"),
-            "33:34: " + getCheckMessage(MSG_KEY, "hidden"),
-            "40:13: " + getCheckMessage(MSG_KEY, "hidden"),
-            "45:18: " + getCheckMessage(MSG_KEY, "hidden"),
-            "50:33: " + getCheckMessage(MSG_KEY, "hidden"),
+            "21:13: " + getCheckMessage(MSG_KEY, "hidden"),
+            "24:39: " + getCheckMessage(MSG_KEY, "hidden"),
+            "31:13: " + getCheckMessage(MSG_KEY, "hidden"),
+            "36:18: " + getCheckMessage(MSG_KEY, "hidden"),
+            "41:33: " + getCheckMessage(MSG_KEY, "hidden"),
+            "52:17: " + getCheckMessage(MSG_KEY, "innerHidden"),
+            "55:26: " + getCheckMessage(MSG_KEY, "innerHidden"),
             "61:17: " + getCheckMessage(MSG_KEY, "innerHidden"),
-            "64:26: " + getCheckMessage(MSG_KEY, "innerHidden"),
-            "70:17: " + getCheckMessage(MSG_KEY, "innerHidden"),
-            "71:17: " + getCheckMessage(MSG_KEY, "hidden"),
-            "76:22: " + getCheckMessage(MSG_KEY, "innerHidden"),
-            "80:22: " + getCheckMessage(MSG_KEY, "hidden"),
-            "85:17: " + getCheckMessage(MSG_KEY, "innerHidden"),
-            "87:17: " + getCheckMessage(MSG_KEY, "hidden"),
-            "93:17: " + getCheckMessage(MSG_KEY, "innerHidden"),
-            "94:17: " + getCheckMessage(MSG_KEY, "hidden"),
-            "99:13: " + getCheckMessage(MSG_KEY, "hidden"),
-            "117:29: " + getCheckMessage(MSG_KEY, "prop"),
-            "123:29: " + getCheckMessage(MSG_KEY, "prop"),
-            "129:29: " + getCheckMessage(MSG_KEY, "prop"),
-            "141:28: " + getCheckMessage(MSG_KEY, "prop"),
-            "155:13: " + getCheckMessage(MSG_KEY, "hidden"),
-            "160:13: " + getCheckMessage(MSG_KEY, "hidden"),
-            "165:13: " + getCheckMessage(MSG_KEY, "hidden"),
-            "169:13: " + getCheckMessage(MSG_KEY, "hidden"),
-            "196:23: " + getCheckMessage(MSG_KEY, "y"),
-            "217:17: " + getCheckMessage(MSG_KEY, "hidden"),
-            "227:22: " + getCheckMessage(MSG_KEY, "hidden"),
-            "234:13: " + getCheckMessage(MSG_KEY, "hidden"),
-            "240:13: " + getCheckMessage(MSG_KEY, "hiddenStatic"),
-            "247:41: " + getCheckMessage(MSG_KEY, "x"),
-            "253:30: " + getCheckMessage(MSG_KEY, "xAxis"),
-            "270:41: " + getCheckMessage(MSG_KEY, "prop"),
-            "284:29: " + getCheckMessage(MSG_KEY, "prop"),
-            "295:42: " + getCheckMessage(MSG_KEY, "prop2"),
-            "307:19: " + getCheckMessage(MSG_KEY, "i"),
+            "62:17: " + getCheckMessage(MSG_KEY, "hidden"),
+            "67:22: " + getCheckMessage(MSG_KEY, "innerHidden"),
+            "71:22: " + getCheckMessage(MSG_KEY, "hidden"),
+            "76:17: " + getCheckMessage(MSG_KEY, "innerHidden"),
+            "78:17: " + getCheckMessage(MSG_KEY, "hidden"),
+            "84:17: " + getCheckMessage(MSG_KEY, "innerHidden"),
+            "85:17: " + getCheckMessage(MSG_KEY, "hidden"),
+            "90:13: " + getCheckMessage(MSG_KEY, "hidden"),
         };
         verifyWithInlineConfigParser(
-                getPath("InputHiddenField2.java"), expected);
+                 getPath("InputHiddenField2Basic.java"), expected);
+    }
+
+    @Test
+    public void testHiddenField2Interface() throws Exception {
+        final String[] expected = {
+        };
+        verifyWithInlineConfigParser(
+                 getPath("InputHiddenField2Interface.java"), expected);
+    }
+
+    @Test
+    public void testHiddenField2Static() throws Exception {
+        final String[] expected = {
+            "22:13: " + getCheckMessage(MSG_KEY, "hidden"),
+            "27:13: " + getCheckMessage(MSG_KEY, "hidden"),
+            "32:13: " + getCheckMessage(MSG_KEY, "hidden"),
+            "36:13: " + getCheckMessage(MSG_KEY, "hidden"),
+            "63:23: " + getCheckMessage(MSG_KEY, "y"),
+        };
+        verifyWithInlineConfigParser(
+                 getPath("InputHiddenField2Static.java"), expected);
+    }
+
+    @Test
+    public void testHiddenField2PropertySetter() throws Exception {
+        final String[] expected = {
+            "21:29: " + getCheckMessage(MSG_KEY, "prop"),
+            "27:29: " + getCheckMessage(MSG_KEY, "prop"),
+            "33:29: " + getCheckMessage(MSG_KEY, "prop"),
+            "45:28: " + getCheckMessage(MSG_KEY, "prop"),
+            "64:41: " + getCheckMessage(MSG_KEY, "prop"),
+            "78:29: " + getCheckMessage(MSG_KEY, "prop"),
+            "89:42: " + getCheckMessage(MSG_KEY, "prop2"),
+        };
+        verifyWithInlineConfigParser(
+                 getPath("InputHiddenField2PropertySetter.java"), expected);
+    }
+
+    @Test
+    public void testHiddenField2Enum() throws Exception {
+        final String[] expected = {
+            "30:17: " + getCheckMessage(MSG_KEY, "hidden"),
+            "40:31: " + getCheckMessage(MSG_KEY, "hidden"),
+            "47:13: " + getCheckMessage(MSG_KEY, "hidden"),
+            "53:13: " + getCheckMessage(MSG_KEY, "hiddenStatic"),
+        };
+        verifyWithInlineConfigParser(
+                 getPath("InputHiddenField2Enum.java"), expected);
+    }
+
+    @Test
+    public void testHiddenField2Misc() throws Exception {
+        final String[] expected = {
+            "19:45: " + getCheckMessage(MSG_KEY, "x"),
+            "24:34: " + getCheckMessage(MSG_KEY, "xAxis"),
+            "31:23: " + getCheckMessage(MSG_KEY, "i"),
+        };
+        verifyWithInlineConfigParser(
+                 getPath("InputHiddenField2Misc.java"), expected);
     }
 
     /** Tests ignoreFormat property. */
