+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/hiddenfield/InputHiddenField2Interface.java
@@ -1,0 +1,21 @@
+/*
+HiddenField
+ignoreFormat = (default)null
+ignoreConstructorParameter = (default)false
+ignoreSetter = (default)false
+setterCanReturnItsClass = (default)false
+ignoreAbstractMethods = (default)false
+tokens = (default)VARIABLE_DEF, PARAMETER_DEF, PATTERN_VARIABLE_DEF, LAMBDA, RECORD_COMPONENT_DEF
+
+
+*/
+
+package com.puppycrawl.tools.checkstyle.checks.coding.hiddenfield;
+
+interface InputHiddenField2Interface
+{
+    public static int notHidden = 0;
+
+    // not a violation
+    public void noShadow(int notHidden);
+}
